PowerWidgetUtil: Remove flashlight as a default button
Remove BUTTON_FLASHLIGHT as a default button in
PowerWidgetUtil::getCurrentButtons so as to remain in sync with the
defaults used by PowerWidget (SystemUI).

Otherwise the power widget will only contain four buttons on first boot
(five including WiMAX), but the flashlight will "mysteriously appear" if
any other button is toggled in the "Widget buttons" settings.